+++ mods/src/api/events.cpp
@@ -1,4 +1,3 @@
-#include <algorithm>
 #include <ranges>
 
 #include <libreborn/patch.h>
@@ -101,22 +100,15 @@ static bool CommandServer__updateAccept_setSocketBlocking_injection(const int fd
     extra_client_data[fd] = ExtraClientData();
     return CommandServer::setSocketBlocking(fd, param_1);
 }
-static bool CommandServer__updateClient_injection(CommandServer__updateClient_t original, CommandServer *self, ConnectedClient &client) {
-    const bool ret = original(self, client);
-    if (!ret) {
-        // Client Disconnected
-        extra_client_data.erase(client.sock);
-    }
-    return ret;
+void api_free_event_data(const int sock) {
+    extra_client_data.erase(sock);
 }
-static void CommandServer__close_injection(CommandServer__close_t original, CommandServer *self) {
-    // Server Shutdown
+void api_free_all_event_data() {
     extra_client_data.clear();
-    original(self);
 }
 
 // Clear All Events
-static void clear_events(const ConnectedClient &client) {
+static void clear_all_events(const ConnectedClient &client) {
     ExtraClientData &data = extra_client_data.at(client.sock);
     if (!api_compat_mode) {
         // Match RJ Bug
@@ -129,9 +121,9 @@ static void clear_events(const ConnectedClient &client) {
 // Clear Events Produced By Given Entity
 template <typename T>
 static void _clear_events(std::vector<T> &data, const int id) {
-    data.erase(std::remove_if(data.begin(), data.end(), [&id](const T &e) {
+    std::erase_if(data, [&id](const T &e) {
         return id == e.owner_id;
-    }), data.end());
+    });
 }
 static void clear_events(const ConnectedClient &client, const int id) {
     ExtraClientData &data = extra_client_data.at(client.sock);
@@ -287,31 +279,29 @@ void _init_api_events() {
     overwrite_calls(Gui_addMessage, Gui_addMessage_injection);
     // Track Connected Clients
     overwrite_call((void *) 0x6bd78, CommandServer_setSocketBlocking, CommandServer__updateAccept_setSocketBlocking_injection);
-    overwrite_calls(CommandServer__updateClient, CommandServer__updateClient_injection);
-    overwrite_calls(CommandServer__close, CommandServer__close_injection);
     // Track Block Hits
     overwrite_calls(GameMode_useItemOn, GameMode_useItemOn_injection);
     overwrite_calls(CreatorMode_useItemOn, CreatorMode_useItemOn_injection);
 }
 
 // Handle Commands
-std::string api_handle_event_command(CommandServer *server, const ConnectedClient &client, const std::string &cmd, std::optional<int> id) {
+std::string api_handle_event_command(CommandServer *server, const ConnectedClient &client, const std::string_view &cmd, const std::optional<int> id) {
     if (cmd == "clear") {
         // Clear Events
         if (id.has_value()) {
             clear_events(client, id.value());
         } else {
-            clear_events(client);
+            clear_all_events(client);
         }
         return CommandServer::NullString;
     } else if (cmd == "chat.posts") {
         // Chat Events
         return get_chat_events(server, client, id);
     } else if (cmd == "block.hits") {
-        // Chat Events
+        // Block Hit Events
         return get_block_hit_events(server, client, id);
     } else if (cmd == "projectile.hits") {
-        // Chat Events
+        // Projectile Events
         return get_projectile_events(server, client, id);
     } else {
         // Invalid Command